+++ src/game/state_checkpoint/model_keyword.rs
@@ -30,6 +30,7 @@ pub(super) enum KeywordSnapshot {
     Mountainwalk,
     Forestwalk,
     Plainswalk,
+    LegendaryLandwalk,
     Islandwalk,
     Swampwalk,
     ProtectionFromWhite,